Изменение реализации создания фильтров
