+++ src/main.js
@@ -1,3 +1,5 @@
+import removeDebounce from './remove-debounce';
+
 import createFilters from './create-filters';
 import createCards from './create-cards';
 
@@ -19,8 +21,41 @@ const deleteExistingFilters = () => {
   }
 };
 
+
+const toggleFilter = (filter) => {
+  const currentActive = filter.parentElement.querySelector(`.main-navigation__item--active`);
+
+  if (currentActive && currentActive !== filter) {
+    currentActive.classList.remove(`main-navigation__item--active`);
+    filter.classList.add(`main-navigation__item--active`);
+
+    createCards(filter.dataset.count);
+  }
+};
+
+const onFilterClick = (evt) => {
+  evt.preventDefault();
+
+  removeDebounce(() => {
+    let target = evt.target;
+
+    while (!target.parentElement.classList.contains(`main-navigation`)) {
+      target = target.parentElement;
+    }
+
+    toggleFilter(target);
+  });
+};
+
+
 deleteExistingFilters();
 createFilters(filtersContainer);
+
+const filters = filtersContainer.querySelectorAll(`.main-navigation__item:not(.main-navigation__item--additional`);
+
+for (const filter of filters) {
+  filter.addEventListener(`click`, onFilterClick);
+}
 
 
 // создание карточек
